Initial commit of gff_to_db
The code compiles and appears to correctly extract gene information from
gff files into a set of 8 related tables.

The correctness has not been tested.

The output files contain fields that should be removed as they are redundant
(eg. names of transcripts and exons). These are output to facilitate testing.